+++ src/app/Components/REGISTER_YOUR_SAMPLES_HERE.cs
@@ -17,6 +17,7 @@ using Components.ComponentsSamples.Selection;
 using Components.ComponentsSamples.Sorting;
 using Components.ComponentsSamples.SyntaxHighlighting;
 using Components.ComponentsSamples.TabView;
+using Components.ComponentsSamples.Tags;
 using Components.ComponentsSamples.Text;
 using Components.ComponentsSamples.TextFields;
 using Components.ComponentsSamples.Tip;
@@ -60,8 +61,8 @@ public static class REGISTER_YOUR_SAMPLES_HERE
             new(SampleType.Components, "Syntax Highlighting", () => new SyntaxHighlightingSamples()),
             new(SampleType.Components, "Amplitude View", () => new AmplitudeViewSamples()),
             new(SampleType.Components, "Text", () => new TextSamples()),
+            new(SampleType.Components, "Tag", () => new TagsSamples()),
             new(SampleType.Components, "TabView", () => new TabViewSamples()),
-