+++ src/library/DIPS.Mobile.UI/API/Camera/ImageCapturing/Views/CameraZoom/ZoomButtons.cs
@@ -171,7 +171,7 @@ internal class ZoomButton : ContentView
         m_label = new Label
         {
             Text = IsDefaultActive ? $"{defaultZoomRatio} ×" : $"{defaultZoomRatio}",
-            TextColor = IsDefaultActive ? DIPS.Mobile.UI.Resources.Colors.Colors.GetColor(ColorName.color_yellow_400) : DIPS.Mobile.UI.Resources.Colors.Colors.GetColor(ColorName.color_text_on_action),
+            TextColor = IsDefaultActive ? Colors.Gold : Colors.White,
             VerticalTextAlignment = TextAlignment.Center,
             HorizontalTextAlignment = TextAlignment.Center,
             Style = Styles.GetLabelStyle(LabelStyle.UI100),