+++ src/library/DIPS.Mobile.UI/Resources/Styles/Tag/TagStyle.cs
@@ -1,0 +1,12 @@
+namespace DIPS.Mobile.UI.Resources.Styles.Tag;
+
+public enum TagStyle
+{
+    None = 0,
+    Default = 1,
+    Danger = 2,
+    Subtle = 3,
+    Success = 4,
+    Warning = 5,
+    Information = 6
+}